+++ GameStudio1Lab2/Assets/Prefabs1/Camera.cs.meta
