+++ GameStudio1Lab2/Assets/Prefabs1/Coin.prefab.meta
