+++ GameStudio1Lab2/Assets/Prefabs1/MarioCollision.cs.meta
