+++ GameStudio1Lab2/Assets/Prefabs1/Star.prefab.meta
