+++ GameStudio1Lab2/Assets/Prefabs1/enemies_1.prefab.meta
