+++ GameStudio1Lab2/Assets/Prefabs1/enemies_3.prefab.meta
