--- GameStudio1Lab2/Assets/Scripts/MarioMovement.cs
@@ -1,36 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class MarioMovement : MonoBehaviour
-{
-    public float JumpForce = 5f;
-    public float moveSpeed = 5f;
-    //public bool isGrounded = false;
-    private Rigidbody2D rb;
-
-    private void Start()
-    {
-        rb = GetComponent<Rigidbody2D>();
-    }
-    void Update()
-    {
-        var movement = Input.GetAxis("Horizontal");
-        transform.position += new Vector3(movement, 0f, 0f) * Time.deltaTime * moveSpeed;
-
-        if (!Mathf.Approximately(0, movement))
-            transform.rotation = movement < 0 ? Quaternion.Euler(0, 180, 0) : Quaternion.identity;
-
-        if (Input.GetButtonDown("Jump") && Mathf.Abs(rb.velocity.y)<0.001f)
-        {
-            rb.AddForce(new Vector2(0, JumpForce), ForceMode2D.Impulse);
-        }
-        
-
-        
-    }
-    
-
-   
-}
